recipes-kernel: Add linux-xlnx-dev bleeding edge recipe
Bleeding edge recipe allows developers to build the latest commit on
linux-xlnx kernel master-next branch. The branch can be modified by the developer if
they wish in the local.conf file by setting KBRANCH variable.
The configuration is taken from meta-xilinx, and is likely from a slightly
older kernel release. The developer should create a new config if there
are new configs or features in the bleeding edge kernel.

Signed-off-by: Sipke Vriend <sipke.vriend@xilinx.com>