recipes-kernel: Move DTB generation to meta-xilinx
oe-core changed the DTB generation to use the linux make system.
http://git.openembedded.org/openembedded-core/commit/?id=72980d5bb465f0640ed451d1ebb9c5d2a210ad0c
This resulted in the kernel make failing as the meta-xilinx contained
dts/dtsi files are not in the kernel tree.
To resolve, move the DTB creation using DTC into meta-xilinx layer.
This will correct the make error and build the DTS files in meta-xilinx.
Additionally the user can set KERNEL_DEVICETREE within local.conf to
build an in-kernel-tree DTS if they desire.

Signed-off-by: Sipke Vriend <sipke.vriend@xilinx.com>