recipes-bsp:u-boot-xlnx: Move patch files out of include file.
Include file is common and should not hold patch files which may no longer
be required in future versions of u-boot. So move the patches currently
required by 2013.01 into the 2013.101 bb file. If for some reasons these
patches are also required in later versions, they should be added to
the new version recipe also.

Signed-off-by: Sipke Vriend <sipke.vriend@xilinx.com>