Don't add the kernel modules to the kernel-base package. This breaks the kernel-modules metapackage which is automatically created.
By adding the kernel module files to the kernel-base package, the kernel-modules metapackage breaks. The metapackage depends on automatically created kernel-module packages (named kernel-module-<modulename>) which are empty if the files are added to the kernel-base package. This causes bitbake to not create the packages for installing later as they wouldn't have any effect anyway. The twist is, that the kernel-modules metapackages depends on these packages. During creation of the rootfs, the process fails, because those packages can't be found.

If modules should be installed in the image, don't use the kernel-base package for that. The kernel-modules packages is the right one.