linux-xlnx: Default to branchless checkout
The linux-xlnx git repository contains tags/commits that are orphaned
and are not located on any branch. To handle this for the future where
existing commits/tags may become orphaned setup the SRC_URI with
nobranch=1 by default except in the case where tracking a branch instead
of a specific commit.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>