Fix up use of '_append +='
The use of '_append' and '+=' together is considered bad practice within
recipes. It is preferred to explicitly add the preceeding spaces where
they are applicable instead of using the '+=' to achieve the same
result.

Change all the uses of '_append +=' to match the recommended pattern,
adding in spaces where explicitly required.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>