u-boot-xlnx_2015.04: Refactor U-Boot SPL generation
Refactor the way the recipe handles depending, configuring and building
SPL with external or internally provided Zynq Platform Init code.

This refactor allows for the handling of all three configuration cases
for how the platform init files are provided, as well as simplifying how
machines define whether SPL/boot.bin can be built.

This change also reduces the need for a seperate SPL recipe and makes it
much simpler to have multiple recipes that can provide this feature.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>