device-tree.bb: Add deprecation warning for 'zynq7-base.dtsi'
The zynq7-base.dtsi include is deprecated and will be removed in the
future. This informs users of its deprecation, in order to encourage
users to move to the 'zynq-7000.dtsi' include.

This include is deprecated in favour of the 'zynq-7000.dtsi' include
which is available and kept up to date in kernel releases as well as the
linux-xlnx vendor tree releases. The device-tree recipe automatically
provides this include (and dependent includes) to allow for custom
device-trees to be built without needing to embed or patch the kernel
source.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>