device-tree: Add include for dummy qspi node
The zynq-7000.dtsi include does not have a qspi node in mainline
kernels. To resolve this issue provide a dummy node with no properties.
The dummy node does not affect the node provided by zynq-7000.dtsi in
other kernels.

This was previously done in the individual device trees, instead this is
now provided by a common include.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>