machine-xilinx-default.inc: Disable 'qemu-usermode' for MicroBlaze
The qemu-usermode machine feature enables the ability for using QEMU
linux-user during the host build. This allows for the distro feature
gobject-introspection-data to use linux-user mode of QEMU to generate
data. But this process is broken on MicroBlaze due to issues with glib
and the MicroBlaze toolchain. This change disables the use of QEMU
linux-user and is a workaround to avoid the execution of the
gobject-introspection-data process which is default enabled in commonly
used distros ('nodistro' and 'poky').

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>