board-common.cfg: Disable 4K sectors on QSPI NOR flash
The 4K sector support makes erasing much slower and prevents using
wear-levelling filesystems like UBI, which needs erase sectors to be
at least 16k in size. Until this situation has been properly resolved
in the kernel, disable the setting.
For a discussion on the topic, and the reason the config exists, see:
https://patchwork.ozlabs.org/patch/380565/

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>
Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>