ep108-zynqmp, zcu102-zynqmp: Add runqemu support
The ep108-zynqmp machine already has support in QEMU however was exposed
as the qemuzynqmp machine. QEMU 2.7 introduced support for the
'xlnx-zcu102' machine model. This allows for runqemu to boot a machine
based on the ep108 or zcu102 hardware. The support added is based on the
configuration that was in qemuzynqmp.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>