u-boot-spl-zynq-init.inc: Enable the FORCE_PS7INIT for new U-Boot
The directory structure in U-Boot for platform init files has changed in
more recent versions of U-Boot. The new naming scheme uses the name of
the device-tree (that is provided in U-Boot) to determine the name of
the directory in board/xilinx/zynq/.

This patch changes how FORCE_PS7INIT functions, such that it is a toggle
instead of requiring the target directory name. FORCE_PS7INIT is changed
to overwrite all the ps7_init_gpl* files in the board/xilinx/zynq/*
directories. This removes the need to specify the target
directory/config and also avoids any issues with the differences in
names between older and newer U-Boot versions.

Additionally this patch adds a [doc] string for the variable.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>