qemu* -> *qemu-*: Rename QEMU machines
Rename the QEMU machines to match the <board>-<design>-<arch>
convention. QEMU only machines no longer need to be named qemu*.

Note for qemuzynq it is special in that it does not model a board, as
such it has no board in the name being just 'qemu-zynq7'.

This updates the README.md to match the new machine names.

Additional update all occurrences of the qemu* machines

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Reviewed-by: Alistair Francis <alistair.francis@xilinx.com>