linux/config: Update and improve config fragments
Update and add new config fragments to enable newer features that are
now in the newer kernels (e.g. Zynq and ZynqMP drivers/features). This
update also attempts to cover ZynqMP configuration that is normally
provided by the ZynqMP defconfig in linux-xlnx, this allows for better
consistency between the Zynq/ZynqMP/MicroBlaze configuration for
linux-xlnx as well as enabling better configuration setup for the
linux-yocto* kernels.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>