u-boot-xlnx: For ZynqMP setup auto-boot for JTAG loading
When booting a ZynqMP machine configured to boot from JTAG, setup U-Boot
to attempt to automatically load the kernel, rootfs and device tree from
default memory locations.

This enables auto-booting for QEMU allowing for a completely automated
boot up when using qemu-xilinx.

Note the patch in this change is not acceptable for upstreaming in its
current form. The upstream maintainer would however accept a similar
change that uses FIT instead of individual components.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Reviewed-by: Alistair Francis <alistair.francis@xilinx.com>