arm-trusted-firmware: Embed Xilinx Release Version into PV
Embed the Xilinx Release Version (aka the 'xilinx-vYYYY.X' tag) into the
PV of the recipe, this makes it easier to determine which version of
Xilinx's ATF is being used.

Additionally change the version string to extend instead of extra data
(switch the '+' to a '-' before 'xilinx...').

This removes the need for the comment on what the SRCREV points at,
instead this is replaced with a XILINX_RELEASE_VERSION variable which
contains the 'vYYYY.X' release string.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>