u-boot-xlnx: Embed Xilinx Release Version into PV
Embed the Xilinx Release Version into the PV of the u-boot-xlnx recipes.
This allows for easier determination of which version of u-boot-xlnx is
in use and avoid any ambiguity when one or more Xilinx Release uses the
same upstream u-boot version.

This removes the need for the comment on what the SRCREV points at,
instead this is replaced with a XILINX_RELEASE_VERSION variable which
contains the 'vYYYY.X' release string.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>