linux-xlnx: Embed Xilinx Release Version in PV and EXTENSION
Embed the Xilinx Release Version into the PV of the linux-xlnx recipes.
This allows for easier determination of which version of linux-xlnx is
in use and avoid any ambiguity when one or more Xilinx Release uses the
same upstream kernel version.

This removes the need for the comment on what the SRCREV points at,
instead this is replaced with a XILINX_RELEASE_VERSION variable which
contains the 'vYYYY.X' release string.

This change sets the LINUX_VERSION_EXTENSION with the release version so
that the kernel is compiled with the version in 'CONFIG_LOCALVERISON',
this results in the kernel having the Xilinx Release Version embedded in
its version which can be seen using uname or during boot:

    $ uname -sr
    Linux 4.6.0-xilinx-v2016.3

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>