glibc: Fix up MicroBlaze pt-vfork compat symbols
A bug has been present in glibc for a while where a symbol is emitted in
libpthread.so which is corrupt. Whilst previously this has not been an
issue due to how binutils has handled it, recent binutils behaviour has
changed enough to make this bug surface as a linker error.

Depending on the host (behaviour and native dependencies) the cross
binutils that is built may emit different values for the corrupt symbol.
This appears as an error similar to the following when linking against
libpthread.so (e.g. librt links against libpthread in glibc):

    invalid string offset 1843814912 >= 4913 for section `.dynstr'

This patch fixes up the symbols emitted for libpthread.so without
breaking any libpthread or libc ABI. See the glibc patch for more
specifics.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>