linux-yocto_4.8: Resolve CONFIG_DEBUG_RODATA for Zynq
This patch resolves a previous commit (ea4ca65aaf) which only mitigated
the issue caused due to CONFIG_DEBUG_RODATA for Zynq.

This issue has been properly fixed in the next version of the kernel
(v4.10) and this change backports the patch to the v4.8 kernel for
linux-yocto(-tiny).

Additionally this change removes the forced disabling of
CONFIG_DEBUG_RODATA allowing for Zynq platforms to use the feature
without issue.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>