qemu-xilinx: Updates for libgcrypt and improved qemu.inc
Update the append to work with the improved qemu.inc. Enable gcrypt
PACKAGECONFIG options.

In order to enable libgcrypt the QEMU configure must be patched to
handle pkg-config support for libgcrypt. This is because upstream
libgcrypt does not provide pkg-config support but OE does. The solution
used works around the need to patch QEMU, by implementing the patch as a
sed replace so that a patch does not need to be used and maintained
across versions.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>