xilinx-platform-init.bbclass: Generic platform-init class
Create a more generic platform init class for use with Zynq 7000, ZynqMP
and or future targets that use the platform initialization code files
configuration setup.

This generic class defines the source files (which are defined per SoC)
and the staging directory to populate with the source files.

This class replaces zynq7-platform-paths.bbclass.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>