platform-init.bb: Update to xilinx-platform-init
Update to use xilinx-platform-init.bbclass. And change to providing the
generic 'virtual/xilinx-platform-init' as opposed to a zynq7 specific
one.

Remove the 'zynq7-platform-init.inc' this include was only useful for
recipes that provide/package only platform-init files. Which means it is
no more useful that the platform-init.bb recipe itself. Layers that want
to provide additional platform-init files can bbappend to the
platform-init recipe.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>