u-boot-spl-zynq-init.inc: Add support for ZynqMP
Update to using the xilinx-platform-init.bbclass and depending on the
'virtual/xilinx-platform-init' provider. This allows for more generic
support of platform-init (between Zynq7, ZynqMP and any future targets).

This change also renames some of the variables used for defining the
source of the platform-init files for SPL. Specifically HAS_PS7INIT is
renamed to HAS_PLATFORM_INIT, and FORCE_PS7INIT is renamed to
FORCE_PLATFORM_INIT. However for compatibility the existing variables
still function as expected.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>