conf/machine/*.conf: Move to QB_NETWORK_DEVICE for QEMU targets
Move the setup of network devices to the new QB_NETWORK_DEVICE variable,
this avoids the need to configure slirp or tap options from the machine
configuration.

Additionally this provides the benefit of letting runqemu handle the
setup of tap and slirp arguments, allowing for it to provide features
such as host-guest port forwarding for slirp.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Tested-by: Alistair Francis <alistair.francis@xilinx.com>