u-boot-xlinx: Workaround to trigger network based jtagboot
Currently the jtagboot cmd fails when jtagmemboot failed. This is
because the jtagmemboot returns false when it fail the itest.w or the
booti failed. Thus the network based jtagboot cmd never gets triggered.

The jtagboot can't be purely base on the itest.w because sometimes
the dhcp command triggers the Image download which has valid image
in the ram. This passes the itest.w and iminfo test but fail the
booti.

This patch triggers the network based jtagboot cmd if jtagmemboot
failed.

Signed-off-by: Jason Wu <jason.hy.wu@gmail.com>
Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>