machine-xilinx-default.inc: Add provider for mesa-gl
With this patch, we switch providers for virtual/mesa and virtual/libgl
to mesa-gl on zynmp.  As a result, the mesa_%.bbappend is no longer
required, and we handle interworking with libmali much better.

As a part of the change, we also move this configuration from
zcu102-zynqmp specific conf to generic conf.

Signed-off-by: Cyril Chemparathy <cyril.chemparathy@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>