device-tree.bbappend: Move use of MACHINE_DEVICETREE to SRC_URI
Remove the use of MACHINE_DEVICETREE for providing sources to
device-tree, instead provide the device tree sources as SRC_URI entires.
Whilst this was already being done automatically by the
MACHINE_DEVICETREE logic it also created a non-standard way of providing
the device tree sources to the recipe.

The setup of SRC_URI is done via the use of a device-tree.bbappend
instead of directly in the device-tree.bb recipe itself. This is for two
reason, separation of the source from the build logic as well as to
provide an example for how a user can extend the device-tree recipe to
provide custom device tree sources in their layer.

This change also moves the paths to the sources around, this is to align
the names of the directories to match the OVERRIDES so that when
expanding the file paths it will search the subdirs based on the
OVERRIDES or machine names.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>