device-tree: Improve, clean up and remove MACHINE_DEVICETREE
This change improves and does a number of clean up to the device-tree
recipe. This includes adding support for building overlays, default
inclusion of kernel device tree sources, updated license value and
general clean ups to the logic.

This change also removes the support for MACHINE_DEVICETREE, and
additionally the now unused xilinx-utils.bbclass.

Note: In order to support overlays dtc version 1.4.3 or newer is needed,
currently in oe-core only 1.4.2 is available.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>