u-boot-xlnx: Add recipe for xilinx-v2017.1 release
Add a recipe for the u-boot-xlnx xilinx-v2017.1 release, this release is
based on v2017.01 of U-Boot. Also add an updated version of the patches.

This patch also moves the FILESEXTRAPATHS prepend to the u-boot-xlnx.inc
due to common use. And a u-boot-xlnx/${XILINX_RELEASE_VERSION} subdir is
added to the list for version specific patches.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>