qemu-xilinx-helper-native: Clean up preferred provider message
Due to this recipe being added there are two providers for
qemu-helper-native, but by default one is never selected. This generates
a message during the build that informs the user to select a provider.
This is unnecessary as the only machines that care about setting the
provider is currently zcu102-zynqmp due to its requirement for
qemu-xilinx all other qemu machines should be using the
qemu-helper-native recipe and non-qemu machines don't use the recipe.

To prevent the message, only make the qemu-xilinx-helper-native recipe
available when it is selected as the preferred provider otherwise skip
it.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Acked-by: Alistair Francis <alistair.francis@xilinx.com>