zcu102-zynqmp: Pin to pre-v2017.1 versions of kernel, u-boot and atf
Pin the versions of linux-xlnx, u-boot-xlnx and arm-trusted-firmware to
the version prior to xilinx-v2017.1. This is due to the added
requirement of pmu-firmware for xilinx-v2017.1. Currently it is not
possible to boot the zcu102-zynqmp machine in QEMU with the newer
versions due to missing the pmu rom as well as the pmu firmware itself.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>