classes/zynqmp-pmu.bbclass: Create a extender class for the ZynqMP PMU
This class is an extender class similar to native/nativesdk/multilib
that should be used with BBCLASSEXTEND.

The purpose of this class is to allow the building of recipes for the
ZynqMP PMU architecture along side building for the primary APU of the
ZynqMP. Which allows for building the PMU firmware itself as well as its
dependencies (e.g. newlib, libgloss).

This class is intended for use with binutils-cross, gcc-cross, newlib,
libgloss and pmu-firmware, but can be used globally. However there are
limitations of this class and functional behaviour beyond the intended
targets is not tested.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Tested-by: Alistair Francis <alistair.francis@xilinx.com>
Tested-by: Koteswararao Nayudu <kotin@xilinx.com>