ep108-zynqmp: Re-enable 'qemuboot' image class
Re-enable the 'qemuboot' image class which is used to generate
qemuboot.conf for the ep108-zynqmp machine. This used to be set by
machine-xilinx-qemu.inc but as of the 'qemuboot-xilinx.bbclass: Rework
qemu-xilinx setup' commit is no longer set by default. The same commit
setup other machines to use the 'qemuboot' image class directly but
skipped this machine.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>