zybo-linux-bd-zynq7: Change gem0 phy to match kernel definition
Change the device tree for the gem0 phy to match that of the kernel
device tree for ZYBO. This removes the compatible string and uses the
phy address of 0 (broadcast). This lets the kernel auto detect the phy
since there are no other devices on the MDIO bus for gem0.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>