README.md: Add paragraph for the 'xilinx' license flag
This paragraph documents the existence of recipes that rely on closed or
otherwise restricted source content that can be used to compile images
for targets.

The content has specific licensing and legal requirements that are
outside the ability or sanity of enforcement in OpenEmbedded or Yocto
and as such is up to each individual user to review and comply with.
This might include EULA, Export Compliance, NDAs, Redistribution, etc.

Generally speaking this content is going to be protected by Xilinx's web
download service which requires an account.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Tested-by: Alistair Francis <alistair.francis@xilinx.com>
Acked-by: Alistair Francis <alistair.francis@xilinx.com>
Acked-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>