kc705-bitstream_2016.3.bb: Update to use the PetaLinux BSP bitstream
This recipe was never updated when the rest of the configuration was
updated to be built for the PetaLinux v2016.3 BSP. This was due to
download access issues which were not resolved.

Instead of pointing at the old bitstream point at the download path on
Xilinx's servers which is restricted by account access, and using the
xilinx-fetch-restricted class to handle erroring properly and providing
user instructions for fetching the BSP file.

This recipe now also manually extracts the 'download.bit' from the BSP
and populates it into a package as well as the deploy directory.

The recipe is now marked with the 'xilinx' license flag due to its use
of content that has licensing/legal requirements.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Tested-by: Alistair Francis <alistair.francis@xilinx.com>
Acked-by: Alistair Francis <alistair.francis@xilinx.com>
Acked-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>