pmu-rom_2017.1.bb: Recipe that deploys pmu-rom.elf from ZCU102 BSP
Add a recipe that extracts the pmu-rom.elf from the PetaLinux ZCU102
BSP. This recipe cannot fetch the BSP, but uses the
xilinx-fetch-restricted class to handle error cases and providing
information on how to manually fetch.

This recipe is marked with the 'xilinx' license flag due to its use of
content that has licensing/legal requirements.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Tested-by: Alistair Francis <alistair.francis@xilinx.com>
Acked-by: Alistair Francis <alistair.francis@xilinx.com>
Acked-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>