pmu-firmware: Create and deploy a binary firmware blob
The SPL based flow needs a raw firmware image instead of an ELF executable. Though
probably any objcopy version can perform this, it's better to have the actual
microblaze tool do the conversion, hence its inclusion in this recipe.

With this addition, it is possible to create a bootloader for the zynqmp
platforms based op opensource tools.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>
Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>