linux/config: Rework kmeta based on kernel-cache merge
A large portion of the kmeta in xilinx-common and xilinx-machine was
upstreamed into the linux-yocto kernel-cache and is no longer needed in
the meta-xilinx layer.

This change removes any fragments that are in the yocto-kernel-cache,
and reworks remaining configs for use with linux-yocto. Compatibly with
linux-xlnx is dropped due to favouring defconfigs.

This includes Zynq and MicroBlaze machine configs (excluding kc705 and
Zybo Linux-BD BSP), board common bsp/ fragments and a number of feature
fragments that are already available in the yocto kernel cache.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>