zcu102-zynqmp: Setup runqemu to boot from an SD image
This sets up the zcu102-zynqmp machine to by default generate a padded
SD card image for which runqemu can use to boot from. This image
includes all the components that are expected from to boot and mirrors
the same requirements that the real board has.

The QEMU args are changed to default to the SD boot mode and only U-Boot
SPL is passed in as a ROM image to boot the QEMU instance.

This setup mimics the boot flow of the physical target where the Boot
ROM loads U-Boot SPL and the PMU Firmware from the boot.bin.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>