feature-microblaze-versions.inc: Rework and expand version conflicts
This change removes the use of the python function to expand the
TUNECONFLICTS definitions for version tunes. Instead this is replaced
with hardcoded definitions. This is preferred due to reduced complexity
of the python code as well as streamlining the include, this is also
done to make the includes align better with expectations of tune
includes in OE-Core.

This change also moves the version conflicts for the reorder tune into
arch-microblaze.

Additional documentation is added to cover details of MicroBlaze CPU
versions, and the change between different version formats.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>