u-boot-spl-zynq-init.inc: Add virtual/boot-bin provider
Add back the PROVIDES for virtual/boot-bin and additionally improve the
logic that enables the provide so that U-Boot can be built without
providing virtual/boot-bin.

In order for the u-boot recipe to provide virtual/boot-bin for Zynq and
ZynqMP targets the selected provider for virtual/boot-bin must be the
current u-boot (e.g. u-boot or u-boot-xlnx) or the provider must be
unset. When u-boot provides virtual/boot-bin it is only enabling the
deployment of boot.bin. In cases where it is not the provider it
forcibly prevents the deployment of boot.bin. This ensures that the
platform-init files are still provided even if SPL is not used for the
boot.bin.

Additionally for ZynqMP if u-boot is not providing the boot.bin do not
depend on virtual/pmu-firmware. Since the boot.bin is what contains the
pmu binary.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>