machine-xilinx-default.inc: Setup MicroBlaze preferred GDB
With the perspective removal of tune-microblaze.inc, the following
preferred GDB version configuration for MicroBlaze needs to be
relocated.

This change also drops the setting of gdb-cross* targets as they were
not setup properly since "gdb-cross-${TARGET_ARCH}" was not specified
correctly. This is also preferred as newer GDB (e.g. 8.0) does support
gdb-cross for debugging (just not gdbserver/gdb).

Ideally the MicroBlaze GDB patches should be updated to support the
newest GDB and or upstreamed. However this setup continues to be
available until that occurs.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>