machine/*-microblaze*: Add linux.bin.ub KERNEL_IMAGETYPE
The tune-microblaze.inc is currently providing this configuration,
however since this is not a tune specific configuration it should be set
by the machine itself. Additionally with the perspective change to
remove tune-microblaze.inc the reliance on this includes configuration
needs to be removed.

Also remove the superfluous '_remove = "device-tree"' for
s3adsp1800-qemu-microblazeeb.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>