device-tree.bb: Change depedency to do_configure task
This patch is required to remove circular depedency while compiling
kernel image simpleImage.mb for MB machines.  We need the task to be
depedent on do_configure because kernel will depend on device-tree
to be available while compiling simpleImage.mb.
The kernel source is available in work-shared/kernel-source after
do_configure task is run, this patch changes the depedency from
do_shared_workdir to do_configure.

Reviewed-by: Nathan Rossi <nathan@nathanrossi.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>