machine-xilinx-overrides: Fix overrides extender
Previously a class to extend MACHINEOVERRIDES was introduced, while this
class was useful to us, it had some fundamental problems, it was
expanding MACHINEOVERRIDES and writing back to it, which caused the
dynamically adjusting mechanism of bitbake to fail.

A previous effort was made to fix this problem, but fixing it caused
another issue regarding when this event was being triggered.

This patch deletes the previously mentioned class in favor of a simple
new function which provides similar functionality, it lets us
modify/extend MACHINEOVERRIDES depending on the SOC_FAMILY and
SOC_VARIANT that were building for, but it wont let us delete anything
from it (because bitbake would also collapse).

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>
Signed-off-by: Vineeth Chowdary Karumanchi <vineethchowz.chowdary@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>